Luxai-s2-Baseline4gymnasium/requirements.txt
absl-py==2.0.0
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.3
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gymnasium==0.29.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
kiwisolver==1.4.5
luxai-s2==3.0.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
munkres==1.1.4
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
PettingZoo==1.12.0
Pillow==10.0.1
platformdirs==3.11.0
ply==3.11
pooch==1.7.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
sip==6.7.11
six==1.16.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
unicodedata2==15.1.0
urllib3==2.0.6
vec-noise==1.1.4
Werkzeug==3.0.0
zipp==3.17.0